Yoast SEO - Gutenberg blocks translation - wpmltriage-2557
Yoast introduced Gutenberg blocks. They are already translatable in ATE. However, when opening the translation in the WP editor, the block will show a "Block contains unexpected or invalid content." message. When clicking the "Attempt recovery" button, the block loses its translation. That's because the block is translated but not the block attributes.