Yoast SEO - Gutenberg blocks translation - wpmltriage-2557
Yoast introduced Gutenberg blocks, which are already translatable in ATE. However, when opening the translation in the WP editor, a "Block contains unexpected or invalid content" message appears. Clicking the "Attempt recovery" button causes the block to lose its translation. This happens because the block is translated, but not the block attributes. The issue is caused by the mismatch between the block's attributes and its content.

Following Pierre's suggestion, I removed classes where possible to simplify the code and broaden its scope.